# backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: 台灣智慧交通路線規劃 API
  version: 1.0.0
  description: |
    🤖 AI-Assisted Development Showcase
    
    這是一個展示現代 Node.js 後端開發技能的專案，
    整合了 AI 輔助開發流程和台灣交通數據處理能力。
    
    ## 主要功能
    - 🗺️ 地點座標處理和驗證
    - 🚇 多種交通工具路線計算
    - 💰 時間與費用成本分析
    - 📊 即時交通數據整合
    
    ## AI 輔助開發特色
    - 智慧型 API 設計
    - 自動化程式碼生成
    - 最佳實踐建議
    
    ## 使用範例
    
    ### 1. 計算路線
    ```bash
    curl -X POST http://localhost:3000/api/routes \
      -H "Content-Type: application/json" \
      -d '{
        "origin": {"latitude": 25.0478, "longitude": 121.5170},
        "destination": {"latitude": 25.0340, "longitude": 121.5645}
      }'
    ```
    
    ### 2. 搜尋地點
    ```bash
    curl "http://localhost:3000/api/locations/search?q=台北車站"
    ```
  contact:
    name: AI-Assisted Development Team
    email: dev@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.example.com
    description: Production server

tags:
  - name: Routes
    description: 路線計算相關 API - 展示演算法設計和效能優化
  - name: Locations
    description: 地點處理相關 API - 展示地理資料處理和搜尋功能
  - name: System
    description: 系統監控相關 API - 展示系統設計和監控能力

paths:
  /health:
    get:
      summary: 系統健康檢查
      description: 檢查 API 服務狀態和系統資訊
      tags: [System]
      responses:
        '200':
          description: 系統正常運行
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "OK"
                  timestamp:
                    type: string
                    format: date-time
                  uptime:
                    type: number
                    description: 系統運行時間（秒）
                    example: 3600
                  environment:
                    type: string
                    example: "development"

  /api:
    get:
      summary: API 資訊
      description: 獲取 API 版本資訊和可用端點列表
      tags: [System]
      responses:
        '200':
          description: API 資訊
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    example: "台灣智慧交通路線規劃 API"
                  version:
                    type: string
                    example: "1.0.0"
                  description:
                    type: string
                    example: "AI-Assisted Development Showcase"
                  endpoints:
                    type: object
                    properties:
                      routes:
                        type: string
                        example: "/api/routes"
                      locations:
                        type: string
                        example: "/api/locations"
                      health:
                        type: string
                        example: "/health"
                      docs:
                        type: string
                        example: "/api-docs"

  /api/routes:
    post:
      summary: 計算路線
      description: 根據起終點座標計算最佳路線組合
      tags: [Routes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - origin
                - destination
              properties:
                origin:
                  $ref: '#/components/schemas/Coordinates'
                destination:
                  $ref: '#/components/schemas/Coordinates'
                preferences:
                  type: object
                  properties:
                    prioritize:
                      type: string
                      enum: [time, cost, transfers, eco]
                      description: 優先考量因素
                    maxWalkingDistance:
                      type: number
                      description: 最大步行距離（公尺）
                    avoidModes:
                      type: array
                      items:
                        type: string
                      description: 避免的交通工具
                    departureTime:
                      type: string
                      format: date-time
                      description: 出發時間
            examples:
              taipei_example:
                summary: 台北市內路線
                value:
                  origin:
                    latitude: 25.0478
                    longitude: 121.5170
                  destination:
                    latitude: 25.0340
                    longitude: 121.5645
                  preferences:
                    prioritize: "time"
                    maxWalkingDistance: 800
      responses:
        '200':
          description: 路線計算成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/routes/stats:
    get:
      summary: 獲取路線服務統計
      description: 獲取路線計算服務的統計資訊
      tags: [Routes]
      responses:
        '200':
          description: 統計資訊
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      cacheSize:
                        type: number
                      cacheHitRate:
                        type: number
                      totalCalculations:
                        type: number
                  timestamp:
                    type: string
                    format: date-time

  /api/locations/reverse-geocode:
    post:
      summary: 反向地理編碼
      description: 將座標轉換為地址資訊
      tags: [Locations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Coordinates'
            example:
              latitude: 25.0478
              longitude: 121.5170
      responses:
        '200':
          description: 地址資訊
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/locations/geocode:
    post:
      summary: 地理編碼
      description: 將地址轉換為座標資訊
      tags: [Locations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - address
              properties:
                address:
                  type: string
                  example: "台北車站"
      responses:
        '200':
          description: 座標資訊列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  /api/locations/search:
    get:
      summary: 搜尋地點
      description: 根據關鍵字搜尋地點
      tags: [Locations]
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
          description: 搜尋關鍵字
          example: "台北"
        - in: query
          name: lat
          schema:
            type: number
          description: 搜尋中心緯度
        - in: query
          name: lng
          schema:
            type: number
          description: 搜尋中心經度
        - in: query
          name: radius
          schema:
            type: number
            default: 5000
          description: 搜尋半徑（公尺）
        - in: query
          name: limit
          schema:
            type: number
            default: 20
          description: 結果數量限制
      responses:
        '200':
          description: 搜尋結果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  /api/locations/nearby-stations:
    post:
      summary: 尋找附近交通站點
      description: 根據座標尋找附近的交通站點
      tags: [Locations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - coordinates
              properties:
                coordinates:
                  $ref: '#/components/schemas/Coordinates'
                maxDistance:
                  type: number
                  default: 1000
                  description: 最大距離（公尺）
                limit:
                  type: number
                  default: 10
                  description: 結果數量限制
      responses:
        '200':
          description: 附近的交通站點
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  /api/locations/cities:
    get:
      summary: 獲取台灣城市列表
      description: 獲取台灣主要城市和熱門景點資訊
      tags: [Locations]
      responses:
        '200':
          description: 城市列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  /api/locations/stations:
    get:
      summary: 獲取交通站點
      description: 根據交通工具類型獲取站點列表
      tags: [Locations]
      parameters:
        - in: query
          name: mode
          schema:
            type: string
            enum: [mrt, bus, hsr, train]
          description: 交通工具類型
      responses:
        '200':
          description: 交通站點列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

components:
  schemas:
    Coordinates:
      type: object
      required: [latitude, longitude]
      description: 地理座標 - 使用 WGS84 座標系統
      properties:
        latitude:
          type: number
          minimum: -90
          maximum: 90
          description: 緯度
          example: 25.0478
        longitude:
          type: number
          minimum: -180
          maximum: 180
          description: 經度
          example: 121.5170

    Location:
      type: object
      required: [coordinates]
      description: 地點資訊
      properties:
        coordinates:
          $ref: '#/components/schemas/Coordinates'
        address:
          type: string
          description: 地址
          example: 台北市中正區北平西路3號
        name:
          type: string
          description: 地點名稱
          example: 台北車站

    Station:
      type: object
      required: [id, name, coordinates, type]
      description: 交通站點資訊
      properties:
        id:
          type: string
          description: 站點 ID
          example: mrt-taipei-main-station
        name:
          type: string
          description: 站點名稱
          example: 台北車站
        coordinates:
          $ref: '#/components/schemas/Coordinates'
        type:
          type: string
          enum: [walk, bus, mrt, train, hsr, youbike, taxi]
          description: 交通工具類型
          example: mrt
        lines:
          type: array
          items:
            type: string
          description: 路線名稱
          example: [淡水信義線, 板南線]
        address:
          type: string
          description: 站點地址
          example: 台北市中正區北平西路3號

    RouteSegment:
      type: object
      required: [mode, from, to, duration, cost, instructions]
      description: 路線段落
      properties:
        mode:
          type: string
          enum: [walk, bus, mrt, train, hsr, youbike, taxi]
          description: 交通工具
          example: mrt
        from:
          $ref: '#/components/schemas/Station'
        to:
          $ref: '#/components/schemas/Station'
        duration:
          type: number
          description: 時間（分鐘）
          example: 25
        cost:
          type: number
          description: 費用（新台幣）
          example: 25
        instructions:
          type: array
          items:
            type: string
          description: 指示說明
          example: [搭乘淡水信義線, 往淡水方向, 在台北101/世貿站下車]
        line:
          type: string
          description: 路線名稱
          example: 淡水信義線
        distance:
          type: number
          description: 距離（公里）
          example: 5.2

    Route:
      type: object
      required: [id, type, totalTime, totalCost, transfers, segments]
      description: 完整路線資訊
      properties:
        id:
          type: string
          description: 路線 ID
          example: route-fastest-001
        type:
          type: string
          description: 路線類型
          example: 最快路線
        totalTime:
          type: number
          description: 總時間（分鐘）
          example: 45
        totalCost:
          type: number
          description: 總費用（新台幣）
          example: 65
        totalDistance:
          type: number
          description: 總距離（公里）
          example: 12.5
        transfers:
          type: number
          description: 轉乘次數
          example: 2
        segments:
          type: array
          items:
            $ref: '#/components/schemas/RouteSegment'
          description: 路線段落
        carbonFootprint:
          type: number
          description: 碳足跡（公克 CO2）
          example: 850

    RouteResponse:
      type: object
      required: [routes, origin, destination, searchTime, alternatives]
      description: 路線查詢回應
      properties:
        routes:
          type: array
          items:
            $ref: '#/components/schemas/Route'
          description: 路線選項
        origin:
          $ref: '#/components/schemas/Location'
        destination:
          $ref: '#/components/schemas/Location'
        searchTime:
          type: string
          format: date-time
          description: 搜尋時間
        alternatives:
          type: number
          description: 替代路線數量
          example: 3

    TaiwanCity:
      type: object
      required: [id, name, englishName, coordinates]
      description: 台灣城市資訊
      properties:
        id:
          type: string
          example: taipei
        name:
          type: string
          example: 台北市
        englishName:
          type: string
          example: Taipei City
        coordinates:
          $ref: '#/components/schemas/Coordinates'
        popularDestinations:
          type: array
          items:
            $ref: '#/components/schemas/Location'
          description: 熱門景點
        transportHubs:
          type: array
          items:
            $ref: '#/components/schemas/Station'
          description: 交通樞紐

    APIError:
      type: object
      required: [error, message, timestamp]
      description: API 錯誤回應
      properties:
        error:
          type: string
          description: 錯誤類型
          example: Invalid coordinates
        message:
          type: string
          description: 錯誤訊息
          example: Latitude must be between -90 and 90
        code:
          type: string
          description: 錯誤代碼
          example: INVALID_COORDINATES
        timestamp:
          type: string
          format: date-time
          description: 錯誤時間
        path:
          type: string
          description: API 路徑
          example: /api/routes
        method:
          type: string
          description: HTTP 方法
          example: POST
        details:
          type: object
          description: 詳細錯誤資訊

    SuccessResponse:
      type: object
      required: [success, data, timestamp]
      description: API 成功回應格式
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: 回應資料
        message:
          type: string
          description: 成功訊息
          example: 操作成功
        timestamp:
          type: string
          format: date-time
          description: 回應時間
        count:
          type: number
          description: 資料筆數（適用於列表回應）

  responses:
    BadRequest:
      description: 請求參數錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            error: Validation Error
            message: Invalid coordinates format
            code: INVALID_REQUEST
            timestamp: '2024-01-01T12:00:00.000Z'
            path: /api/routes
            method: POST

    NotFound:
      description: 資源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            error: Not Found
            message: No routes found for the given coordinates
            code: NO_ROUTES_FOUND
            timestamp: '2024-01-01T12:00:00.000Z'

    InternalServerError:
      description: 伺服器內部錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            error: Internal Server Error
            message: An unexpected error occurred
            code: INTERNAL_ERROR
            timestamp: '2024-01-01T12:00:00.000Z'

    TooManyRequests:
      description: 請求過於頻繁
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            error: Too Many Requests
            message: Rate limit exceeded. Please try again later.
            code: RATE_LIMIT_EXCEEDED
            timestamp: '2024-01-01T12:00:00.000Z'